HBX-1836: Move the ant tasks from the core hibernate tools into the hibernate-tools-ant project - Move the new ant tasks to package 'org.hibernate.tool.ant.fresh' for now
