--- ant/src/main/java/org/hibernate/tool/ant/ExportDdlTask.java
@@ -1,5 +1,0 @@
-package org.hibernate.tool.ant;
-
-public class ExportDdlTask {
-
-}
