+++ ant/src/main/java/org/hibernate/tool/ant/fresh/ExportCfgTask.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.ant;
+package org.hibernate.tool.ant.fresh;
 
 import java.io.File;
 import java.nio.file.Path;
