+++ ant/src/main/java/org/hibernate/tool/ant/fresh/ExportDdlTask.java
@@ -1,0 +1,5 @@
+package org.hibernate.tool.ant.fresh;
+
+public class ExportDdlTask {
+
+}
