+++ ant/src/test/java/org/hibernate/tool/ant/fresh/HibernateToolTaskTest.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.ant;
+package org.hibernate.tool.ant.fresh;
 
 import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertNull;
@@ -7,6 +7,10 @@ import static org.junit.jupiter.api.Assertions.assertSame;
 import java.util.Properties;
 
 import org.hibernate.boot.Metadata;
+import org.hibernate.tool.ant.fresh.ExportCfgTask;
+import org.hibernate.tool.ant.fresh.ExportDdlTask;
+import org.hibernate.tool.ant.fresh.HibernateToolTask;
+import org.hibernate.tool.ant.fresh.MetadataTask;
 import org.hibernate.tool.api.export.ExporterConstants;
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
 import org.junit.jupiter.api.Test;
